move 'os' key to explicit stage to avoid matrix
Having the 'os' key at the top level causes Travis to create a build matrix
with an implicit 'test' stage that runs before explicit 'test' stages.
And that implicit stage runs the default script for C++ projects, which is
`./configure && make && make test`.

But Firefox doesn't build with those commands, so we can't use that default.
We thus need to not use a build matrix at all, and instead specify 'os' in an
explicit 'test' stage whose 'script' property we control.